requirements.txt
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-daq==0.5.0
dash-leaflet==0.1.23
dash-extensions==1.0.0
pandas==1.5.3
geopandas==0.12.2
plotly==5.13.1
joblib==1.2.0
fpdf2==2.7.4
requests==2.31.0 